pyproject.toml
[project]
name = "anki-api"
version = "0.1.0"
description = "Agent-assisted Anki flashcard generation system"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "click>=8.1.0",
    "python-docx>=1.2.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "websockets>=12.0",
    "claude-agent-sdk>=0.1.19",
    "youtube-transcript-api>=1.0.0",
]

[project.scripts]
anki-api = "src.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "PTH",  # flake8-use-pathlib
    "RUF",  # ruff-specific rules
]
ignore = [
    "E501",   # line too long (formatter handles)
    "B008",   # function call in default argument (FastAPI Depends)
    "B904",   # raise from (noisy in FastAPI exception handlers)
    "RUF001", # ambiguous unicode (intentional emojis)
    "RUF013", # implicit Optional (verbose, handled by type checker)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ty]
python-version = "3.11"
